split out twitter fixer function and fix T.A.L.A.T.
