+++ events/messages/messageCreate.js
@@ -1,6 +1,8 @@
 import { clientId, defaultArray, rareArray, rareFrequency, positiveArray, negativeArray, neutralArray, twitterFixEnabled } from '../../configVars.js';
 import dataLog from '../../logging/dataLog.js';
 
+import twitterFixer from './utilities/twitterFixer.js';
+
 // var Sentiment = require('sentiment');
 // var sentiment = new Sentiment();
 
@@ -9,36 +11,42 @@ const name = 'messageCreate';
 let limit = 0;
 let lastReplyTimestamp = null;
 let ms = (Math.random() * 60000) + 60000; // CUTTING EDGE AI ANTI EXPLOIT TECHNOLOGY
-// let ms = 10000;
 
 let configuredLimit = 2;
-
-let once = false;
 
 //******* RESPONSE FUNCTIONS *******//
 
 // takeALook()
 //	selects an image and sends a reply containing the link
 //	return: response (string)
+
 const takeALook = () => {
 	let nowTime = Date.now();
 
+	let response = '';
+
+	// If there has not been a last reply or it has been long enough since the last reply, reset counter
 	if (lastReplyTimestamp == null || Math.floor((nowTime - lastReplyTimestamp)) >= ms) {
-		timestamp = null
+		lastReplyTimestamp = null;
 		limit = 0;
 	}
 
+	//if the spam counter hasnt hit the limit
 	if (limit < configuredLimit) {
 		//continue to response
+
+	//else if the counter is at the limit
 	} else if (limit == (configuredLimit)) {
 
+		//store the time of the last response and count it
 		lastReplyTimestamp = Date.now();
 		limit++;
 
-		return "No spam!"; // https://i.imgur.com/kAClxb0.png = spam picture url lol
+		response = 'No spam!'; // https://i.imgur.com/kAClxb0.png = spam picture url lol
 	}
 
-	let imgLink = "";
+
+	let imgLink = '';
 
 	//New calculation
 	let diceRoll = Math.random();
@@ -49,10 +57,12 @@ const takeALook = () => {
 		imgLink = defaultArray[defaultArray.length * Math.random() | 0]
 	}
 
+	response.length < 1 ? response = imgLink : null;
+
 	limit++;
 	lastReplyTimestamp = Date.now();
 
-	return imgLink;
+	return response;
 }
 
 // fortuneTeller()
@@ -70,32 +80,6 @@ const fortuneTeller = (rawMessage) => {
 	return (combinedResponses[Math.floor(Math.random() * combinedResponses.length)]);
 }
 
-// twitterFixer(str)
-//  reply with a vx twitter link if a non-vx twitter link is posted
-//	note: will use the last link in the message
-//  return: response (string)
-const twitterFixer = (messageContents) => {
-	
-	const msgAry = messageContents.split(' ');
-	let reply = '';
-
-	msgAry.forEach(word => {
-
-		let cleanWord = word.replace(/[<>]/g, '');
-
-		if(cleanWord.startsWith('https://x.com')){
-			if(messageContents.includes('dd') ||
-				messageContents.includes('dixbot') ||
-				messageContents.includes('fix')){
-
-				reply = ("fixed link: " + cleanWord.replace('https://x.com/', 'https://vxtwitter.com/'));
-			}
-		}
-
-	});
-
-	return reply;
-}
 
 // emojiDetector 
 //  logs instances of each emojis in the message
@@ -115,18 +99,17 @@ const execute = (message) => {
 
 	//******* INCOMING MESSAGE PROCESSING *******//
 
-	let response = "";
+	let response = '';
 
 	if (!message.author.bot && !(message.author.id === clientId)) {
 
 		// check every message for emojis
 		emojiDetector(message);
 
-		// Split message content to check each word for trigger keywords
-		const sentence = message.content.split(' ');
-
 		// Strip incoming message for comparison		
 		const contentStripped = message.content.toLowerCase().replace(/[^a-zA-Z0-9]/g, "");
+
+		console.log(contentStripped);
 
 		// If there's a twitter link to fix, do that
 		if (twitterFixEnabled) {
@@ -135,7 +118,10 @@ const execute = (message) => {
 				response = twitFixReply;
 			}
 			//if not, check if there's a take a look at this to reply to
-		} else if (contentStripped.includes("takealookatthis")) {
+		} 
+		
+		if (contentStripped.includes("takealookatthis")) {
+			console.log("taking a look")
 			response = takeALook();
 
 			//if not, check if there is a fortune teller request to reply to
@@ -147,7 +133,6 @@ const execute = (message) => {
 		if (response.length > 0) {
 			// console.log("<@" + message.author.username + '>: ' + message.cleanContent);
 			message.reply(response);
-			return;
 		}
 
 		// If it wasnt a dixbot keyword, log the message for later bot training purposes
